Add user roles and authorization with Pundit

+++ app/controllers/wikis_controller.rb
@@ -7,6 +7,8 @@ class WikisController < ApplicationController
 
   def show
     @wiki = Wiki.find(params[:id])
+    
+    authorize @wiki
   end
 
   def new
@@ -19,17 +21,21 @@ class WikisController < ApplicationController
     @wiki.body = params[:wiki][:body]
     @wiki.user = current_user
     
+    authorize @wiki
+    
     if @wiki.save
       flash[:notice] = "Wiki was saved."
       redirect_to @wiki
     else
-      flash.now[:alert] = "There was an error saving the wiki. Please try again."
+      flash.now[:danger] = "There was an error saving the wiki. Please try again."
       render :new
     end
   end
 
   def edit
     @wiki = Wiki.find(params[:id])
+    
+    authorize @wiki
   end
   
   def update
@@ -37,11 +43,13 @@ class WikisController < ApplicationController
     @wiki.title = params[:wiki][:title]
     @wiki.body = params[:wiki][:body]
     
+    authorize @wiki
+    
     if @wiki.save
       flash[:notice] = "Wiki was updated succesfully."
       redirect_to @wiki
     else
-      flash.now[:alert] = "There was an error updating the wiki. Please try again."
+      flash.now[:danger] = "There was an error updating the wiki. Please try again."
       render :edit
     end
   end
@@ -49,11 +57,13 @@ class WikisController < ApplicationController
   def destroy
     @wiki = Wiki.find(params[:id])
     
+    authorize @wiki
+    
     if @wiki.destroy
       flash[:notice] = "Wiki deleted succesfully."
       redirect_to wikis_path
     else
-      flash.now[:alert] = "There was an error deleting the wiki. Please try again."
+      flash.now[:danger] = "There was an error deleting the wiki. Please try again."
       render :show
     end
   end
